refactor(html): tree construction "text"
